+++ mongo-api/src/main/java/com/linjun/service/ApiCodeService.java
@@ -13,12 +13,12 @@ import java.util.Map;
  **/
 @Service
 public interface ApiCodeService {
-    CodeEntity queryObject(Integer id);
+    CodeEntity queryObject(Long id);
     List<CodeEntity> queryList(Map<String,Object> map);
     int queryTotal(Map<String,Object> map);
     void save(CodeEntity codeEntity);
     void update(CodeEntity codeEntity);
-    void delete(Integer id);
-    void deleteBatch(Integer[] ids);
+    void delete(Long id);
+    void deleteBatch(Long[] ids);
 
 }
