+++ mongo-api/src/main/java/com/linjun/service/ApiRentMessageService.java
@@ -13,11 +13,11 @@ import java.util.Map;
  **/
 @Service
 public interface ApiRentMessageService {
-    RentMessageEntity queryObject(Integer id);
+    RentMessageEntity queryObject(Long id);
     List<RentMessageEntity> queryList(Map<String,Object> map);
     int queryTotal(Map<String,Object> map);
     void save(RentMessageEntity rentMessageEntity);
     void update(RentMessageEntity rentMessageEntity);
-    void delete(Integer id);
-    void deleteBatch(Integer[] ids);
+    void delete(Long id);
+    void deleteBatch(Long[] ids);
 }
