+++ mongo-api/src/main/java/com/linjun/service/ApiSignService.java
@@ -13,12 +13,12 @@ import java.util.Map;
  **/
 @Service
 public interface ApiSignService {
-    SignEntity queryObject(Integer id);
+    SignEntity queryObject(Long id);
     List<SignEntity> queryList(Map<String,Object> map);
     int queryTotal(Map<String,Object> map);
     void save(SignEntity signEntity);
     void update(SignEntity signEntity);
-    void delete(Integer id);
-    void deleteBatch(Integer[] ids);
+    void delete(Long id);
+    void deleteBatch(Long[] ids);
 
 }
