+++ mongo-api/src/main/java/com/linjun/service/ApiSkillCategoryService.java
@@ -13,12 +13,12 @@ import java.util.Map;
  **/
 @Service
 public interface ApiSkillCategoryService {
-    SkillCategoryEntity queryObject(Integer id);
+    SkillCategoryEntity queryObject(Long id);
     List<SkillCategoryEntity> queryList(Map<String,Object> map);
     int queryTotal(Map<String,Object> map);
     void save(SkillCategoryEntity skillCategoryEntity);
     void update(SkillCategoryEntity skillCategoryEntity);
-    void delete(Integer id);
-        void deleteBatch(Integer[] ids);
+    void delete(Long id);
+        void deleteBatch(Long[] ids);
 
 }
