+++ mongo-api/src/main/java/com/linjun/service/ApiSkillEvaluateService.java
@@ -14,12 +14,12 @@ import java.util.Map;
  **/
 @Service
 public interface ApiSkillEvaluateService {
-    SkillEvaluateEntity queryObject(Integer id);
+    SkillEvaluateEntity queryObject(Long id);
     List<SkillEvaluateEntity> queryList(Map<String,Object> map);
     int queryTotal(Map<String,Object> map);
     void save(SkillEvaluateEntity skillEvaluateEntity);
     void update(SkillEvaluateEntity skillEvaluateEntity);
-    void delete(Integer id);
-    void deleteBatch(Integer[] ids);
+    void delete(Long id);
+    void deleteBatch(Long[] ids);
 
 }
